requirements.txt
streamlit
opencv-python
tensorflow
numpy
pillow
